10-12-2023 18:35

+++ src/Pages/AdminRol.js
@@ -5,7 +5,7 @@ import './../styles/AdminRol.css';
 function AdminRol() {
   const [rol, setRol] = useState([]);
   const [newRol, setNewRol] = useState("");
-  const [isEditing, setIsEditing] = useState(false);
+  const [editedRoles, setEditedRoles] = useState({});
   const [editedRol, setEditedRol] = useState("");
   const [editedRoleId, setEditedRoleId] = useState("");
 
@@ -28,7 +28,7 @@ function AdminRol() {
 
   const handleNewRolSubmit = async (e) => {
     e.preventDefault();
-  
+
     try {
       const response = await fetch('https://entreraices-production.up.railway.app/api/rol/create', {
         method: 'POST',
@@ -37,17 +37,17 @@ function AdminRol() {
         },
         body: JSON.stringify({ rol: newRol }),
       });
-  
+
       if (!response.ok) {
         throw new Error('Error al agregar un nuevo rol');
       }
-  
+
       const updatedRol = await response.json();
       if (updatedRol && updatedRol.categories) {
         const updatedCategoriesArray = Object.values(updatedRol.categories);
-  
+
         setRol((prevRol) => [...prevRol, ...updatedCategoriesArray]);
-  
+
         console.log('Nuevo rol agregado con éxito');
         setNewRol("");
       } else {
@@ -57,10 +57,9 @@ function AdminRol() {
       console.error('Error al agregar un nuevo rol:', error);
     }
   };
-  
 
   const handleEditClick = (rolId, rolName) => {
-    setIsEditing(true);
+    setEditedRoles({ ...editedRoles, [rolId]: true });
     setEditedRol(rolName);
     setEditedRoleId(rolId);
   };
@@ -85,7 +84,7 @@ function AdminRol() {
         )
       );
 
-      setIsEditing(false);
+      setEditedRoles({});
       setEditedRol("");
       setEditedRoleId("");
     } catch (error) {
@@ -93,8 +92,20 @@ function AdminRol() {
     }
   };
 
-  const handleDeleteClick = async (rolId) => {
-    try {
+  const handleCancelEdit = (rolId) => {
+    const updatedEditedRoles = { ...editedRoles };
+    delete updatedEditedRoles[rolId];
+    setEditedRoles(updatedEditedRoles);
+    setEditedRol("");
+    setEditedRoleId("");
+  };
+
+ 
+const handleDeleteClick = async (rolId) => {
+  try {
+    const confirmDelete = window.confirm("¿Estás seguro de que deseas eliminar este rol?");
+    
+    if (confirmDelete) {
       const response = await fetch('https://entreraices-production.up.railway.app/api/rol/delete', {
         method: 'DELETE',
         headers: {
@@ -102,21 +113,21 @@ function AdminRol() {
         },
         body: JSON.stringify({ rol_id: rolId }),
       });
-  
+
       if (!response.ok) {
         throw new Error('Error al eliminar el rol');
       }
-  
+
       const rolToDelete = rol.find((rol) => rol.id_rol === rolId);
       console.log('Datos a eliminar:', rolToDelete);
-  
+
       setRol((prevRol) => prevRol.filter((rol) => rol.id_rol !== rolId));
       console.log('Rol eliminado con éxito:', rolId);
-    } catch (error) {
-      console.error('Error al eliminar el rol:', error);
     }
-  };
-  
+  } catch (error) {
+    console.error('Error al eliminar el rol:', error);
+  }
+};
 
   return (
     <div className="admin-rol-container">
@@ -149,41 +160,51 @@ function AdminRol() {
           </tr>
         </thead>
         <tbody>
-          {rol.map((rol) => (
-            <tr key={rol.id_rol} className="admin-rol-even-row">
+          {rol.map((rolItem) => (
+            <tr key={rolItem.id_rol} className="admin-rol-even-row">
               <td className="admin-rol-cell">
-                {isEditing && editedRoleId === rol.id_rol ? (
+                {editedRoles[rolItem.id_rol] ? (
                   <input
                     type="text"
                     value={editedRol}
                     onChange={(e) => setEditedRol(e.target.value)}
                   />
                 ) : (
-                  rol.rol
+                  rolItem.rol
                 )}
               </td>
               <td className="admin-rol-cell">
-                {isEditing && editedRoleId === rol.id_rol ? (
-                  <button
-                    className="admin-rol-action-button"
-                    onClick={handleSaveEdit}
-                  >
-                    Guardar
-                  </button>
+                {editedRoles[rolItem.id_rol] ? (
+                  <>
+                    <button
+                      className="admin-rol-action-button"
+                      onClick={handleSaveEdit}
+                    >
+                      Guardar
+                    </button>
+                    <button
+                      className="admin-rol-action-button"
+                      onClick={() => handleCancelEdit(rolItem.id_rol)}
+                    >
+                      Cancelar
+                    </button>
+                  </>
                 ) : (
-                  <button
-                    className="admin-rol-action-button"
-                    onClick={() => handleEditClick(rol.id_rol, rol.rol)}
-                  >
-                    Modificar
-                  </button>
+                  <>
+                    <button
+                      className="admin-rol-action-button"
+                      onClick={() => handleEditClick(rolItem.id_rol, rolItem.rol)}
+                    >
+                      Modificar
+                    </button>
+                    <button
+                      className="admin-rol-action-button"
+                      onClick={() => handleDeleteClick(rolItem.id_rol)}
+                    >
+                      Eliminar
+                    </button>
+                  </>
                 )}
-                <button
-                  className="admin-rol-action-button"
-                  onClick={() => handleDeleteClick(rol.id_rol)}
-                >
-                  Eliminar
-                </button>
               </td>
             </tr>
           ))}
